requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0

# Data collection and processing
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# OpenAI and fine-tuning
openai>=1.0.0

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Web interface
streamlit>=1.28.0

# Orchestration and scheduling
schedule>=1.2.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality and linting
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Security scanning
bandit>=1.7.0

# Utilities
aiohttp>=3.8.0
